Update F_keleres.php w/ Commander Clarification
Had a game recently that ran into this fun rules hole. The action card that you use to unlock the Keleres commander does not itself trigger the Keleres hero (due to weirdness with how exactly triggers work in TI). 
So (ignoring interactions with Fleet Logistics and the Master Plan action card) the turn you unlock her you do not get additional actions.
I figured this may be worth adding here. I tried my best to find the most concise way to phrase this here.
IDK if the interactions with Fleet Logistics and Master Plan should be clarified as well...